:recycle: refactor: MemberQueryService 패키지 변경
