+++ src/main/java/org/withtime/be/withtimebe/domain/member/service/query/MemberQueryService.java
@@ -1,4 +1,4 @@
-package org.withtime.be.withtimebe.domain.member.service;
+package org.withtime.be.withtimebe.domain.member.service.query;
 
 import org.withtime.be.withtimebe.domain.member.entity.Member;
